Add files via upload
New environment would probably need be made, to make this functional. If your running vs code it's easy just open a new empty file and let the ide create the environment for you; via Python: create environment. Then just pip install necessary libraries. It's not finished but will continue to work on it